Reorganize FlowStart folder
