+++ front/app/containers/Authentication/steps/FlowStart/_components/AdminSignInLink.tsx
@@ -8,7 +8,7 @@ import { getAzureConfig } from 'api/id_methods/utils';
 
 import { useIntl } from 'utils/cl-intl';
 
-import messages from './messages';
+import messages from '../messages';
 
 const StyledA = styled.a`
   font-size: ${fontSizes.base}px;
