+++ front/app/containers/Authentication/steps/FlowStart/_components/VerificationWarning.tsx
@@ -11,7 +11,7 @@ import Warning from 'components/UI/Warning';
 
 import { useIntl } from 'utils/cl-intl';
 
-import messages from './messages';
+import messages from '../messages';
 
 // The Warning component has a teal (light blue) background; override the icon
 // colour to a dark blue shield to signal that verification is required.
