--- front/app/containers/Authentication/steps/FlowStart/types.ts
@@ -1,3 +1,0 @@
-export interface FormValues {
-  email: string;
-}
